Move logic to replace batch out of async 'submit_batch' function
It's possible that when we call the async 'submit_batch' function
execution yields to the runtime immediately before the 'take' occurs to
replace the batch. Then, two concurrent 'submit_batch' calls could
result in a batch that is never submitted.

I'm not sure at which point execution of the async 'submit_batch' yields
to the runtime, but this change shouldn't hurt, at least.