Add TlogTile wrapper around Tile to implement tlog-tiles spec
The Tile struct supports the tile format used by the Go checksum library
(from which it was ported), but the tlog-tiles spec makes a few changes.
Specifically, it fixes the height at 8 (so each tile has 256 elements),
and uses the path element "entries" for data tiles.

With this change, we can remove custom path logic from the static_ct_api
crate, although we still need to use the path element "data".

I don't love the "data_path_opt" field in tiles, but this seems like an
OK way to support both "data" and "entries" as the data path element.