Double-check that R2 object uploads are successful
As a measure to prevent #84, immediately confirm that we can fetch the
correct objects from R2 after they are uploaded.

I was having trouble getting the sentry integration working, so in the
meantime update the Cleaner to just check that each full level-0 tile
exists. This will stop progress indefinitely to allow us to detect if
more tiles go missing.