Fix handling of fatal sequencing error, closes #84 (#88)
* Fix handling of fatal sequencing error, closes #84

This bug was introduced in f10355f when trying to reload the sequencer
state right away when a fatal sequencing error occurs. If the reloading
failed, the old state would not be cleared, as all the calling function
does is log the error message.

In this fix, we force the DO to reload the state the next time the alarm
or fetch handlers run.

* Cleanup

- Remove unused dependencies (cargo machete)
- Rename UNHASHED_DATA_TILE_LEVEL -> AUX
- Remove a few unnecessary clones and collects