bugfix: ensure accepted root is included in chain_fingerprints
Due to a bug introduced in fb4f44d, the `certificate_chain` field of
data tiles may not include the accepted root certificate, which is a
violation of the RFC6962 requirement (inherited by static-ct-api) that
"[t]he final certificate MUST be a root certificate accepted by the
log."

This bug is triggered by add-[pre-]chain requests where the submitted
chain does not include a log-accepted root certificate. Instead of
storing the full validated chain including the accepted root
certificate, the stored `certificate_chain` only includes the
intermediate certificates submitted in the request.

The bug does not impact chain validation, so no entries were accepted
into the log that would have otherwise been accepted in the absence of
the bug.

It is also possible to recover from the bug.  As the `certificate_chain`
field of data tiles is not authenticated in the tree, it should be
possible to update data tiles to append the trusted root in cases when
it is missing.

Thanks to Andrew Ayer for spotting and reporting this bug.