Extend Cleaner DO to repair data tiles, fixes #99
Some data tiles do not include the inferred root in the
(unauthenticated) chain_fingerprints field. The issue has been fixed for
new tiles that are written, but we need to repair old tiles.

To do this, iterate over all data tiles in the log and check if the last
hash in the chain_fingerprints field corresponds to an accepted root. If
not, find an an appropriate trusted root and append the hash to the
chain. Afterwards, back up the old tile and write the new tile to the
data path.

Once this fix has been applied to all active logs and the cleaner
catches up to the target size, we can remove this code.