Parallelize and add counts of changed entries/tiles
This significantly speeds up each individual invocation of the cleaner
(previously, it took ~5 minutes, but now it takes a few seconds). The
downside is that we're not being as optimal when counting subrequests.
If we reserve 4 subrequests per tile cleaning, we can only clean 249
tiles per invocation, or roughly 4 tiles/second. We can make up for this
by reducing the cleaning interval to 10s for the dev log.