Add metrics to measure cost of extra check after put
As part of the investigation of #84, 4ac28d9 added an extra GET request after each R2 PUT operation to double-check that entries are indeed persisted (something guaranteed by R2, but we wanted to rule out that possibility). If the cost is essentially zero, there's no harm in keeping it, but otherwise remove it to reduce sequencing latency. This commit adds metrics to help make that determination.