Keep only a single copy of roots per worker script, fixes #110
In 36bb0fc, we switched from loading roots once per Worker script (with
LazyLock) to loading roots dynamically for each fetch request, since we
need an async operation to fetch roots from KV, and LazyLock doesn't
support async initialization. This means each add-chain request has to
keep a separate copy of the roots in the Worker memoery. Combined with
ad6ced9 which caused the dynamically loaded roots to be kept in memory
across the asynchronous call to the sequencer, this resulted in the
memory issues observed in #110.

Use the runtime-agnostic tokio::sync::LazyCell to restore the previous
behavior of loading roots only once for each Worker script.