Add option to stop chain validation when a trusted path is found
This option allows the Merkle Tree CA to validate cross-signed bootstrap
chains, where the final cert in the chain is not necessarily trusted.
This option is disabled for CT where every cert in the submitted chain
needs to be part of the validated path.

Also enforce path length constraints in validate_chain_lax. RFC 6962
doesn't mention invalid path length constraints as a reason to reject a
certificate chain, but it is required in RFC 9162. Seems fine to reject
these chains in CT. For MTC we do want to enforce this check.