mtc_worker: Make the witness a proper MTC cosigner
The Worker currently produces two cosignatures for each checkpoint. The
first is from the CA itself. The second mocks a second co-signer that
relying parties may include as per their own policy, such as a witness
or mirror. Normally these parties wouldn't be co-located with the CA,
but mocking them is useful for testing purposes.

The cosigner is a plain TLOG cosigner. That is, it doesn't have an
MTC-style cosigner ID and doesn't sign subtrees with the format defined
in Section 5.4.1 of the MTC spec.

This commit replaces the second cosigner with a proper MTC cosigner. Its
ID is determined by the config file. Each log has been assigned a unique
OID from 1.3.6.1.4.1.44363.47, the arc we've designated for MTC.

Note: This is a breaking change for the existing logs, since the second
cosigner has changed. In particular, attempts to add new entries will
result in `TlogError::MissingVerifierSignature`.