mtc_worker: Remove extra cosigner
Currently mtc_worker loads two cosigners whenever it mints a checkpoint:
one for the MTCA itself; and another for a mocked witness. However the
mocked witness isn't a proper MTC cosigner, as it doesn't format the
message to be signed properly.

To avoid confusing users, just remove the witness altogether.