mtc_worker: Update OIDs for id-alg-mtcProof, id-rdna-trustAnchorID
The relative OID for our cosigner and log IDs start are in the arc
"44363.48", which we have allocated for MTC experimentaion. At the
moment, id-rdna-trustAnchorID ends with "44363.47". First, this is the
wrong arc for MTC experimentation. Second, id-rdna-trustAnchorID should
probably not be on Cloudflare's arc, since others may end up with
boostrap MTC logs as well.

Change id-rdna-trustAnchorID from "1.3.6.1.4.1.44363.47" to
"1.3.6.1.4.1".

Similarly, change id-alg-mtcProof from "1.3.6.1.4.1.44363.47.0" to
"1.3.6.1.4.1.44363.48.0".