Implement `Display` for `RelativeOid`
For the purposes of computing the tlog key ID, the log ID is used as the
key name. Add a method for encoding the arcs as a string so that we can
properly construct the key name.

Implementation note: Currently we only keep around the DER encoding of
the arcs. We could decode the arcs from the DER blob, but it's a bit
more convenient to just keep the arcs around in the data structure.