mtc_api: Properly compute the key name and ID when signing subtrees
MTC changes the way the key name and ID are computed compared to TLOG.
First, the key name is constructed from the log ID rather than the API
endpoint. Second, the key ID does not incorporate the public key as it
does in TLOG.

This is a breaking change: Because the signatures won't verify,
mtc_worker won't be able to open older check points.