mtc_api: Sign checkpoints instead of subtrees
MTC cosigners sign two types of notes: one for checkpoints and another
for subtrees. Apart from the notes having a slightly different format,
we need to construct a separate key ID for each case. To accommodate
this, we need to be able to construct an instance of the cosigenr based
on the type of note we're signing.

Replace `MTCSubtreeCosigner` with a more general `MtcCosigner` struct.
The idea is that the constructor will define which type of message is
being signed (a checkpoint or a subtree) and construct the key ID
accordingly. At the moment, we don't actually use the MTC cosigner to
sign subtrees; for now, we just use it to sign checkpoints. Replace the
current constructor `new()` with one that that constructs a checkpoint
signer, `new_checkpoint()`.

Likewise for the note verifier.

This is a breaking change: Because the signatures won't verify,
mtc_worker won't be able to open older check points.